Improve connection fail fast logic
For a very long time, we've had logic to quickly close all clients when
the a server could not be connected to.

This wasn't working correctly in all cases, one of those was when an
auth_user was set (and no auth_db). This was reported in #989.

While looking closer into that issue I realized it also wouldn't handle
cases where first connecting worked but then after a while it doesn't
anymore, because it relied only on `wait_for_welcome`. It's not strictly
necessary to kill all waiting connections when creating new server
connections is not possible anymore, as long as there are still working
server connections. But if those don't exist either, then we're
effectively in the same state as complete refresh (except that we have
cached the welcome message).

Fixes #989
Fixes #985

Co-Authored-By: Anna Khanova <annakhanova@neon.tech>